// src/types/payroll.types.ts
import { Employee } from "./employee.types";

export enum DeductionType {
  INCOME_TAX = "INCOME_TAX", // 소득세
  NATIONAL_PENSION = "NATIONAL_PENSION", // 국민연금
  HEALTH_INSURANCE = "HEALTH_INSURANCE", // 건강보험
  EMPLOYMENT_INSURANCE = "EMPLOYMENT_INSURANCE", // 고용보험
  LONG_TERM_CARE_INSURANCE = "LONG_TERM_CARE_INSURANCE", // 장기요양보험
  OTHER = "OTHER", // 기타
}

export interface PayrollDeduction {
  id: string;
  type: DeductionType;
  amount: number;
}

export interface Payroll {
  id: string;
  employee: Employee;
  employeeId: string;
  paymentMonth: string;
  baseSalary: number;
  bonus: number;
  totalDeductions: number;
  netPay: number;
  deductions: PayrollDeduction[];
  createdAt: Date;
  updatedAt: Date;
}

export interface UpsertSalaryDto {
  baseSalary: number;
  effectiveDate: string;
}